feat: ensure pure builds are run by make
Note that this loops over an array of "build modifiers", which is just
a list of strings that get appended to the name of the environment,
which also happens to be the name of the build. That's getting a bit
overloaded, but not a priority at the moment.

This also adds a list of "fixme" builds, which are skipped for now
because they don't work. There's enough in-flight at the moment that I'd
prefer to get the working pure builds in and come back to the problem
builds once the release blockers are out of the way.