feat: use vendoring step in Rust pure build
This adds an impure build to vendor dependencies that can later be used
in the Rust pure build. Creating the vendored dependencies is easy, but
using the dependencies in the build requires creating a config file for
Cargo to read. Cargo reads _some_ environment variables at runtime,
just not these variables in particular.