ruby: Fix pure build
By splitting the dependency fetching into an impure build and
referencing the cache directory from the pure build.

This avoids using or modifying unversioned config and cache from the
current working directory so that the build is as reproducible as
possible.

It took a while to find the right Bundler options to achieve this from
these docs:

- https://bundler.io/v2.5/man/bundle-config.1.html
- https://bundler.io/v2.5/man/bundle-cache.1.html
- https://bundler.io/v2.5/man/bundle-install.1.html

Even so, the `bundle cache --no-install` command appears to always copy
the gems twice, so I'm manually deleting the non-cache version.

I'm unsure how much of this was hard because the project structure
doesn't match what you'd get from `bundle gem <name>`, which create its
own binstub and gemspec, but I don't want to make too many changes at
once.