go: Use GOMODCACHE for pure build
So that we don't write any state to the current working directory which
may or may not behave the same on subsequent builds.

The build was previously missing a `go mod vendor` step, which is a good
example of the problem because the untracked dir probably existed in a
working directory at some point.

The `-modcacherw` flag is required to prevent subsequent builds from
getting file permission errors until #3216:

    Building hello-deps2-0.0.0 in local mode
    rm: cannot remove '/tmp/store_38fd7277e1af1cdb58a232c2457b3e28-hello-deps2-0.0.0/golang.org/x/text@v0.0.0-20170915032832-14c0d48ead0c/cmd/gotext/message.go': Permission denied
    rm: cannot remove '/tmp/store_38fd7277e1af1cdb58a232c2457b3e28-hello-deps2-0.0.0/golang.org/x/text@v0.0.0-20170915032832-14c0d48ead0c/cmd/gotext/doc.go': Permission denied
    …
    rm: cannot remove '/tmp/store_38fd7277e1af1cdb58a232c2457b3e28-hello-deps2-0.0.0/rsc.io/quote@v1.5.2/quote.go': Permission denied
    rm: cannot remove '/tmp/store_38fd7277e1af1cdb58a232c2457b3e28-hello-deps2-0.0.0/rsc.io/quote@v1.5.2/README.md': Permission denied
    make: *** [/nix/store/ydih3cmpy0zg1jihxvqs784vdi90qwiw-package-builder-1.0.0/libexec/flox-build.mk:569: hello_deps2_local_build] Error 1
    ❌ ERROR: Build failed with status: exit status: 2