test: Use nix run github:flox/flox
Test against HEAD of the `main` branch so that we can exercise new
fixes, features, and catch regressions.

This is needed right now to unblock the PRs for pure manifest builds
which depend on a fix that has just been merged in `flox/flox`. We can
figure out later on how to gate unreleased feature changes to the
examples.

It doesn't quite feel right defining the default in two places but I'm
slightly uncomfortable with only declaring it in one and it matches the
existing style.

Switched to a plain Nix installer since it's confusing to have an unused
Flox binary available, it's slightly faster, and we can apply the
changes to `nix.conf` more easily.