tests(ci): Run on macOS too
So that we get feedback from both standard operating systems that we
distribute Flox for. We don't need to test multi-arch here though
because the unit tests in `flox/flox` will have already done so
sufficiently.

Using ARM64 for macOS because it's significantly faster. I haven't
changed Linux from x86 at the moment though, until we find a need to:

- https://docs.github.com/en/actions/how-tos/writing-workflows/choosing-what-your-workflow-does/running-variations-of-jobs-in-a-workflow#standard-github-hosted-runners-for-public-repositories