fix(nodejs): Quote shell script
The `$@` was being evaluated eagerly when the build script was running
which resulted in it being dropped from the rendered script.

Unfortunately we can't single quote the whole string because we need
`$out` to be evaluated and the escape also needs escaping otherwise we
get:

    ❌ ERROR: Failed to parse manifest:

    invalid escape sequence
    expected `b`, `f`, `n`, `r`, `t`, `u`, `U`, `\`, `"`

Converted to a HEREDOC at the same time to make it easier to read.