feat: Nix build example for Go
References:

- https://nixos.org/manual/nixpkgs/stable/#sec-language-go
- https://flox.dev/docs/concepts/nix-expression-builds/#example-building-a-third-party-project

It can't generate a fixed output derivation with the correct hashes from
`go.sum` alone so we have to provide a `vendorHash`. This seems fine
while we only have a single dependency and CI will fail the build if
incorrect.

An alternative would be to vendor the dependency, which I'm not super
keen on because it's additional cruft in the repo and easy to get
out-of-sync.